优化代码样式
